+++ tests/plugins/removeformat/plugin.js
@@ -129,12 +129,20 @@ bender.test(
 		bender.assert.isInnerHtmlMatching( expected, data );
 	},
 
+	// #12311
+	'test remove format for cite element': function() {
+		this.editorBot.setHtmlWithSelection( '<p>[foo <cite>bar</cite> baz]</p>' );
+		this.editor.execCommand( 'removeFormat' );
+
+		assert.areEqual( '<p>foo bar baz</p>', this.editor.getData(), 'Cite element should be removed.' );
+	},
+
 	'test editor#addRemoveFormatFilter': function() {
 		bender.editorBot.create( {
 			name: 'test_editor2',
 			config: { allowedContent: true }
 		}, function( bot ) {
-			bot.setHtmlWithSelection( '[<p><span style="color:red">foo</span> <b>bar</b></p>]' );
+			bot.setHtmlWithSelection( '<p>[<span style="color:red">foo</span> <b>bar</b>]</p>' );
 
 			bot.editor.addRemoveFormatFilter( function( element ) {
 				return !element.is( 'b' ); // Don't remove 'b' elements.